Refactor(bookmark): Create BookmarkController (#879)
